requirements.txt
paddlepaddle==2.1.0
python-doctr==0.2.1
paddleocr==2.0.6
pytesseract==0.3.7
numpy==1.19.2
opencv-python==4.2.0.32
six==1.15.0
